Move MemoryAllocator to separate header file.
